+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Feat: allow multiple calls for `Builder.state` and `Builder.error`.
   - Feat: add `ITrimmedLexer/IReadonlyTrimmedLexer`.
   - Feat: `Builder.append` can accept actions with different data bindings in one call.
+  - Feat: `Builder.ignore` won't update the token data bindings.
   - Fix: `Lexer.trimStart` shouldn't collect non-muted errors.
 - Parser
   - **_Breaking Change_**: lexer will be trimmed after every parse. This should improve the performance.